Update property names to plural
